fix(fullscreen): PR fixes and project standards
* move to forEach
* remove anonymous func in favor of fat arrow
* move logic to its on file
* add index responsible for only exporting (like useMouse)
